pyproject.toml
[project]
name = "openpi-franka"
version = "0.1.0"
description = "Franka + Robotiq + RealSense interface for openpi"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "numpy>=1.22.4,<2.0.0",
    "pyrealsense2>=2.56.5.9235",
    "franky-control>=1.1.3",
    "pyrobotiqgripper>=2.0.5",
    "rpyc>=6.0.2",
    "opencv-python>=4.10.0.84",
    "tqdm",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["franka_control"]
